+++ src/main/java/janggi/controller/JanggiFlow.java
@@ -6,12 +6,12 @@ import janggi.domain.board.Board;
 import janggi.domain.piece.Piece;
 import janggi.domain.state.GameContext;
 import janggi.exception.JanggiException;
-import janggi.strategy.ArrangementStrategy;
-import janggi.strategy.ArrangementStrategyFactory;
-import janggi.strategy.BoardAssembler;
-import janggi.strategy.IntersectionInitializer;
-import janggi.strategy.PalaceIntersectionInitializer;
-import janggi.strategy.StrategyLabel;
+import janggi.domain.strategy.ArrangementStrategy;
+import janggi.domain.strategy.ArrangementStrategyFactory;
+import janggi.domain.strategy.BoardAssembler;
+import janggi.domain.strategy.IntersectionInitializer;
+import janggi.domain.strategy.PalaceIntersectionInitializer;
+import janggi.domain.strategy.StrategyLabel;
 import janggi.view.ApplicationView;
 import janggi.view.resolver.PieceViewResolver;
 import janggi.view.resolver.SideViewResolver;
