+++ src/main/java/janggi/domain/strategy/ArrangementStrategyFactory.java
@@ -1,4 +1,4 @@
-package janggi.strategy;
+package janggi.domain.strategy;
 
 import janggi.domain.Side;
 import java.util.EnumMap;
