+++ src/main/java/janggi/domain/strategy/MaSangSangMa.java
@@ -1,15 +1,15 @@
-package janggi.strategy;
+package janggi.domain.strategy;
 
 import janggi.domain.Side;
 import janggi.domain.piece.Ma;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.Sang;
 
-public class SangMaMaSang implements ArrangementStrategy {
+public class MaSangSangMa implements ArrangementStrategy {
 
     private final Side side;
 
-    public SangMaMaSang(Side side) {
+    public MaSangSangMa(Side side) {
         this.side = side;
     }
 
@@ -17,14 +17,14 @@ public class SangMaMaSang implements ArrangementStrategy {
     public void place(Piece[][] board) {
         int boardMaxLength = board.length;
         int row = calculateRow(boardMaxLength, side);
-        board[row][1] = new Sang(side);
-        board[row][2] = new Ma(side);
-        board[row][6] = new Ma(side);
-        board[row][7] = new Sang(side);
+        board[row][1] = new Ma(side);
+        board[row][2] = new Sang(side);
+        board[row][6] = new Sang(side);
+        board[row][7] = new Ma(side);
     }
 
     @Override
     public ArrangementStrategy create(Side side) {
-        return new SangMaMaSang(side);
+        return new MaSangSangMa(side);
     }
 }
