+++ src/main/java/janggi/domain/strategy/StrategyLabel.java
@@ -1,4 +1,4 @@
-package janggi.strategy;
+package janggi.domain.strategy;
 
 import java.util.Arrays;
