+++ src/test/java/janggi/domain/board/BoardTest.java
@@ -7,10 +7,10 @@ import janggi.domain.Side;
 import janggi.domain.piece.EmptyPiece;
 import janggi.domain.piece.Piece;
 import janggi.exception.JanggiException;
-import janggi.strategy.ArrangementStrategy;
-import janggi.strategy.BoardAssembler;
-import janggi.strategy.PalaceIntersectionInitializer;
-import janggi.strategy.IntersectionInitializer;
+import janggi.domain.strategy.ArrangementStrategy;
+import janggi.domain.strategy.BoardAssembler;
+import janggi.domain.strategy.PalaceIntersectionInitializer;
+import janggi.domain.strategy.IntersectionInitializer;
 import janggi.support.TestArrangementStrategy;
 import janggi.support.TestPiece;
 import java.util.List;
