+++ src/test/java/janggi/strategy/PalaceIntersectionInitializerTest.java
@@ -5,6 +5,8 @@ import static janggi.domain.rule.route.Direction.*;
 
 import janggi.domain.Intersection;
 import janggi.domain.location.Vector;
+import janggi.domain.strategy.IntersectionInitializer;
+import janggi.domain.strategy.PalaceIntersectionInitializer;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
